company create,databasemody->biginteger, messages.blade.php

+++ database/migrations/2023_05_27_083440_create_companies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('companies', function (Blueprint $table) {
             $table->id();
             $table->string("company_name");
-            $table->integer("taxNumber");
+            $table->bigInteger("taxNumber");
             $table->integer("phone_number");
             $table->string("company_email");
             $table->timestamps();
